+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.5] - 2025-07-20
+
+### Added
+- More endpoints supported by default.
+
+## [5.0.4] - 2025-06-24
+
+### Changed
+- CloudSettings will now reuse the HTTP session from the query client for sync clients.
+
+### Fixed
+
+- CloudSettings now cached by authority (schema, host and port) instead of full URL
+
 ## [5.0.3] - 2025-05-04
 
 ### Fixed
@@ -12,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Ingest from dataframe
 - Dependency issues
 - Typing issues
+
+
 
 ## [5.0.2] - 2025-05-30